adding registration excel generation
