+++ api/src/events/event.ts
@@ -1,0 +1,10 @@
+export interface IEvent {
+  _id: number,
+  title: string,
+  date: Date,
+  address: string,
+  link: string,
+  disciplines: {},
+  registrationPeriodStartDate: Date,
+  registrationPeriodEndDate: Date
+}
