--- api/src/events/events.event.interface.ts
@@ -1,10 +1,0 @@
-export interface IEventData {
-    _id: number,
-    title: string,
-    date: Date,
-    address: string,
-    link: string,
-    disciplines: {},
-    registrationPeriodStartDate: Date,
-    registrationPeriodEndDate: Date
-}
