+++ api/src/members/member.ts
@@ -1,0 +1,3 @@
+export interface IMember {
+  name: string;
+}
