--- api/src/members/members.member.interface.ts
@@ -1,3 +1,0 @@
-export interface IMemberData {
-    name: string;
-}
