--- api/src/registrations/registrations.registration.interface.ts
@@ -1,11 +1,0 @@
-export enum Gender { female, male }
-
-export interface IRegistrationData {
-    eventId: string,
-    name: string,
-    gender: Gender,
-    email: string,
-    birthYear: number,
-    recaptcha: any;
-    disciplines: Array<any>;
-}
