+++ api/src/server.ts
@@ -1,10 +1,12 @@
-import * as express from "express";    
+import * as express from "express";
 import * as bodyParser from "body-parser";
 import * as mongoose from 'mongoose';
 import * as jwt from 'express-jwt';
 import * as cors from 'cors';
 import { Routes } from './routes';
 import { Auth0Configuration } from './auth0.config';
+
+var json2xls = require('json2xls');
 
 // Initialize mongoose
 mongoose.connect('mongodb://localhost/myathleticsclub');
@@ -15,8 +17,8 @@ let app: express.Application = express();
 // Initialize Auth0
 var configuration = new Auth0Configuration();
 var jwtCheck = jwt({
-    secret: new Buffer(configuration.secretKey, 'base64'),
-    audience: configuration.audience
+  secret: new Buffer(configuration.secretKey, 'base64'),
+  audience: configuration.audience
 });
 
 // Configure our app to use bodyParser(it let us get the json data from a POST)
@@ -26,6 +28,9 @@ app.use(bodyParser.json());
 // Allow CORS
 app.use(cors());
 
+// Enable json2xls
+app.use(json2xls.middleware);
+
 // Init routes
 let routes = new Routes(jwtCheck);
 routes.registerRoutes(app);
@@ -33,5 +38,5 @@ routes.registerRoutes(app);
 // Start listening
 let port = process.env.PORT || 8889;
 app.listen(port, () => {
-    console.log('Server listening at: http://localhost:' + port);
+  console.log('Server listening at: http://localhost:' + port);
 });
