+++ api/src/sysevents/sysevent.ts
@@ -1,0 +1,3 @@
+export interface ISysEvent {
+  title: string;
+}
