+++ api/src/sysevents/sysevents.routes.ts
@@ -1,29 +1,29 @@
 import * as express from "express";
 import * as jwt from 'express-jwt';
-import { SysEventsRepository } from './sysevents.repository';
+import { SysEventsService } from './sysevents.service';
 
-export class SysEventRoutes {
+export class SysEventsRoutes {
 
-    syseventsRepository: SysEventsRepository;
-    jwtCheck: any;
+  syseventsService: SysEventsService;
+  jwtCheck: any;
 
-    constructor(jwtCheck: jwt.RequestHandler) {
-        this.syseventsRepository = new SysEventsRepository();
-        this.jwtCheck = jwtCheck;
-    }
+  constructor(jwtCheck: jwt.RequestHandler) {
+    this.syseventsService = new SysEventsService();
+    this.jwtCheck = jwtCheck;
+  }
 
-    registerRoutes(app: express.Application) {
-        // Add new system event
-        app.post('/sysevents', (req, res) => {
-            console.log('adding sysevent to repository');
-            let sysevent = req.body;
-            console.log('sysevent', sysevent);
-            this.syseventsRepository.add(sysevent).then(sysevent => {
-                res.status(201).json(sysevent);
-            }).catch(err => {
-                console.error('ERROR', err);
-                res.status(500).send(err.toString());
-            });
-        });
-    }
+  registerRoutes(app: express.Application) {
+    // Add new system event
+    app.post('/sysevents', (req, res) => {
+      console.log('adding sysevent to repository');
+      let sysevent = req.body;
+      console.log('sysevent', sysevent);
+      this.syseventsService.add(sysevent).then(sysevent => {
+        res.status(201).json(sysevent);
+      }).catch(err => {
+        console.error('ERROR', err);
+        res.status(500).send(err.toString());
+      });
+    });
+  }
 }
