--- api/src/sysevents/sysevents.sysevent.interface.ts
@@ -1,3 +1,0 @@
-export interface ISysEventData {
-    title: string;
-}
