+++ frontend/src/events/event.edit.html
@@ -92,6 +92,7 @@
 </div>
 
 <h2>Deltagere</h2>
+<p><a class="larger-text" ng-href="{{ vm.excelDownloadUrl }}">Download deltagerliste</a></p>
 <div class="row" ng-show="vm.registrations.length > 0">
     <div class="col-sm-6" ng-repeat="registration in vm.registrations">
         <div class="panel panel-default">
