+++ frontend/src/events/events.service.ts
@@ -157,34 +157,38 @@ module events {
 
         getAllDisciplines() {
             return [
-                { id: '40m', name: '40m' },
-                { id: '60m', name: '60m' },
-                { id: '100m', name: '100m' },
-                { id: '200m', name: '200m' },
-                { id: '300m', name: '300m' },
-                { id: '400m', name: '400m' },
-                { id: '800m', name: '800m' },
-                { id: '1500m', name: '1500m' },
-                { id: '3000m', name: '3000m' },
-                { id: '5000m', name: '5000m' },
-                { id: '2000mObstacle', name: '2000m forhindringsløb' },
-                { id: '40mHurdles', name: '40m hæk' },
-                { id: '60mHurdles', name: '60m hæk' },
-                { id: '80mHurdles', name: '80m hæk' },
-                { id: '100mHurdles', name: '100m hæk' },
-                { id: '110mHurdles', name: '110m hæk' },
-                { id: '400mHurdles', name: '400m hæk' },
-                { id: '4x100m', name: '4 x 100m stafet' },
-                { id: '4x400m', name: '4 x 400m stafet' },
-                { id: 'longjump', name: 'længdespring' },
-                { id: 'highjump', name: 'højdespring' },
-                { id: 'shotput', name: 'kuglestød' },
-                { id: 'javelin', name: 'spydkast' },
-                { id: 'discusthrow', name: 'diskoskast' },
-                { id: 'hammerthrow', name: 'hammerkast' },
-                { id: 'weightthrow', name: 'vægtkast' },
-                { id: 'triplejump', name: 'trespring' },
-                { id: 'polevault', name: 'stangspring' },
+                { id: '40m', name: '40m', short: '40' },
+                { id: '60m', name: '60m', short: '60' },
+                { id: '80m', name: '80m', short: '80' },
+                { id: '100m', name: '100m', short: '100' },
+                { id: '200m', name: '200m', short: '200' },
+                { id: '400m', name: '400m', short: '400' },
+                { id: '600m', name: '600m', short: '600' },
+                { id: '800m', name: '800m', short: '800' },
+                { id: '1000m', name: '1000m', short: '1000' },
+                { id: '1500m', name: '1500m', short: '1500' },
+                { id: '3000m', name: '3000m', short: '3000' },
+                { id: '5000m', name: '5000m', short: '5000' },
+                { id: '40mHurdles', name: '40m hæk', short: '40H' },
+                { id: '60mHurdles', name: '60m hæk', short: '60H' },
+                { id: '80mHurdles', name: '80m hæk', short: '80H'},
+                { id: '100mHurdles', name: '100m hæk', short: '100H' },
+                { id: '110mHurdles', name: '110m hæk', short: '110H' },
+                { id: '400mHurdles', name: '400m hæk', short: '400H' },
+                { id: '4x40m', name: '4 x 40m stafet', short: '4x40' },
+                { id: '4x60m', name: '4 x 60m stafet', short: '4x60' },
+                { id: '4x80m', name: '4 x 80m stafet', short: '4x80' },
+                { id: '4x100m', name: '4 x 100m stafet', short: '4x100' },
+                { id: '4x400m', name: '4 x 400m stafet', short: '4x400' },
+                { id: 'longjump', name: 'længdespring', short: 'LÆ' },
+                { id: 'highjump', name: 'højdespring', short: 'HØ' },
+                { id: 'shotput', name: 'kuglestød', short: 'KU' },
+                { id: 'javelin', name: 'spydkast', short: 'SP' },
+                { id: 'discusthrow', name: 'diskoskast', short: 'DI' },
+                { id: 'hammerthrow', name: 'hammerkast', short: 'HA' },
+                { id: 'weightthrow', name: 'vægtkast', short: 'VÆ' },
+                { id: 'triplejump', name: 'trespring', short: 'TR' },
+                { id: 'polevault', name: 'stangspring', short: 'ST' },
             ];
         }
     }
